+++ internal/cli/orchestra_run_runtime.go
@@ -1,0 +1,10 @@
+package cli
+
+import "github.com/insajin/autopus-adk/pkg/orchestra"
+
+var (
+	orchestraRunLoadConfig      = loadOrchestraConfig
+	orchestraRunBuildProviders  = buildProviderConfigs
+	orchestraRunBackendFactory  = orchestra.NewSubprocessBackendImpl
+	orchestraRunExecutePipeline = orchestra.RunSubprocessPipeline
+)